tests/wpt/tests/webnn/permissions-policy/webnn-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that webnn is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissions-policy-features">
<link rel="help" href="https://www.w3.org/TR/webnn/#permissions-policy">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
test(() => {
    assert_in_array('webnn', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise webnn.');
</script>
